% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when_is_lunch.R
\name{when_is_lunch}
\alias{when_is_lunch}
\title{An R function for telling you when lunch time is}
\usage{
when_is_lunch(location = "CSAN")
}
\arguments{
\item{location}{Specify the location for which you want to know when lunch will be.}
}
\value{
A string telling you when lunch time is. The function also prints this to the console.
}
\description{
In case you wanted to use R to figure out when everyone else will be eating lunch, you can use this function.
}
\examples{
when_is_lunch()
when_is_lunch(location = "CSAN")
when_is_lunch(location = "Sweden")
when_is_lunch(location = "Australia")
}
\keyword{location}
\keyword{lunch}
\keyword{time}
